requirements.txt
amqp==2.5.2
appnope==0.1.0
astroid==2.3.3
backcall==0.2.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
cycler==0.10.0
decorator==4.4.2
hurry.filesize==0.9
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imutils==0.5.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
katna==0.4.1.4
kiwisolver==1.2.0
kombu==4.6.8
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
moviepy==1.0.3
networkx==2.5
numpy==1.19.2
opencv-contrib-python==4.4.0.44
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
proglog==0.1.9
prompt-toolkit==3.0.8
ptyprocess==0.6.0
Pygments==2.7.1
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.2
redis==3.5.0
requests==2.24.0
rope==0.16.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
six==1.14.0
threadpoolctl==2.1.0
tifffile==2020.10.1
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.4
typed-ast==1.4.1
urllib3==1.25.11
vine==1.3.0
watchdog==0.10.2
wcwidth==0.2.5
wrapt==1.11.2
